requirements.txt
africastalking==1.2.9
aiofiles==24.1.0
anyio==4.9.0
asgiref==3.8.1
async-property==0.2.2
Authlib==1.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coverage==7.8.0
cryptography==44.0.2
deprecation==2.1.0
Django==5.2
django-environ==0.12.0
django-graphql-jwt==0.3.4
django-oidc-auth==0.0.11
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
drf-oidc-auth==3.0.0
exceptiongroup==1.2.2
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
josepy==2.0.0
jwcrypto==1.5.6
mozilla-django-oidc==4.0.1
packaging==24.2
phonenumbers==9.0.3
pluggy==1.5.0
promise==2.3
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.17
pyjwkest==0.6.2
PyJWT==2.10.1
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-keycloak==5.5.0
requests==2.32.3
requests-toolbelt==1.0.0
schema==0.7.7
six==1.17.0
sniffio==1.3.1
South==1.0.2
sqlparse==0.5.3
text-unidecode==1.3
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.4.0
gunicorn
psycopg2-binary
django
whitenoise
dj-database-url==2.3.0
